投稿を「Card.vue」にまとめて管理
