+++ plugins/parts.js
@@ -1,0 +1,9 @@
+import Vue from 'vue'
+
+import Card from '../components/Card'
+
+Vue.mixin({
+  components: {
+    Card,
+  }
+})
